+++ CHANGELOG.md
@@ -2,6 +2,10 @@
 All notable changes to this project will be documented in this file. This
 project adheres to [Semantic Versioning](http://semver.org/).
 
+## HEAD
+### Added
+- Added SHA256 checksum calculation to rake_helper.rb.
+
 ## Version 1.0.0 - 2015-04-06
 This release is a complete re-write. None of the code from the previous
 release was retained.