+++ CHANGELOG.md
@@ -2,7 +2,7 @@
 All notable changes to this project will be documented in this file. This
 project adheres to [Semantic Versioning](http://semver.org/).
 
-## UNRELEASED
+## Version 1.1.0 - 2015-04-xx
 ### Added
 - Added SHA256 checksum calculation to rake_helper.rb.
 - Added `:use_random_org` key to Passphrase#inspect.